# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIREST Sistema de Pacientes",
        "description": "Documentación de la API para gestionar el sistema pacientes",
        "contact": {
            "name": "Gustavo Galarza- Daniel Mamani",
            "email": "https://github.com/GustavoGalarza/sistema-pos.git"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/medicos": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Obtener lista de médicos",
                "description": "Devuelve todos los médicos registrados (Requiere autenticación)",
                "operationId": "879398b190c580259412a77e767a59f8",
                "responses": {
                    "200": {
                        "description": "Lista de médicos obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Dr. Juan Pérez"
                                            },
                                            "especialidad": {
                                                "type": "string",
                                                "example": "Cardiología"
                                            },
                                            "telefono": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "juan.perez@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Agregar un médico",
                "description": "Crea un nuevo médico (Requiere autenticación)",
                "operationId": "b86744ba3552836c275a4422431ef97a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "especialidad",
                                    "email",
                                    "telefono"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Dr. Juan Pérez"
                                    },
                                    "especialidad": {
                                        "type": "string",
                                        "example": "Cardiología"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Médico agregado satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Médico agregado satisfactoriamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/medicosAll": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Obtener todos los médicos",
                "description": "Devuelve todos los médicos registrados (Requiere autenticación)",
                "operationId": "af67e80100c215cce427a084afce8997",
                "responses": {
                    "200": {
                        "description": "Lista de médicos obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Dr. Juan Pérez"
                                            },
                                            "especialidad": {
                                                "type": "string",
                                                "example": "Cardiología"
                                            },
                                            "telefono": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "juan.perez@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}